+++ database/School.Database/Tables/dbo/Subscription.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [Subscription]
 (
 	StudentId bigint not null,
-	ClassId tinyint not null,
+	ClassId int not null,
 	CONSTRAINT PK_Sub PRIMARY KEY(StudentId,ClassId),
 	CONSTRAINT FK_Subscription_Student FOREIGN KEY(StudentId)
 	REFERENCES [Student],
